Include Region in PhysicalResourceId
To properly handle Update ops that change the Region, make the
PhysicalResourceId include the Region.

This means if the Region is changed, the PhysicalResourceId will also
change. CloudFormation catches this change and issues a Delete request
on the old PhysicalResourceId (with old properties), which will result
in our removing the domain from SES in the old region (immediately after
the Update that added it to the new region).

**BREAKING CHANGE:** CF exposes the PhysicalResourceId as the `!Ref`,
and we had documented this as being the Domain. The new format should
be considered undocumented. Code that uses `!Ref MySESResource` should
change to `!GetAtt MySESResource.Domain` instead.